Images and BookingProducts are ready

+++ backend/api/Repositories/BookingServiceRepository.cs
@@ -10,7 +10,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace api.Repositories
 {
-    public class BookingServiceRepository : IBookingRepository
+    public class BookingServiceRepository : IBookingServiceRepository
     {
         private readonly ApplicationDBContext _dbContext;
         private readonly IServiceRepository _serviceRepository;
@@ -22,25 +22,28 @@ namespace api.Repositories
         public async Task<Booking?> CreateAsync(Booking bookingModel, string RequesterId)
         {
             var serviceModel = await _serviceRepository.GetByIdAsync((int)bookingModel.ServiceId);
-            if (serviceModel == null || serviceModel.OwnerId != RequesterId || bookingModel.RenterId == serviceModel.OwnerId) {
+            if (serviceModel == null || serviceModel.OwnerId != RequesterId || bookingModel.RenterId == serviceModel.OwnerId)
+            {
                 return null;
             }
 
             await _dbContext.Bookings.AddAsync(bookingModel);
             await _dbContext.SaveChangesAsync();
             return bookingModel;
-            
+
         }
 
         public async Task<Booking?> DeleteAsync(int bookingId, string RequesterId)
         {
-           var bookingModel = await _dbContext.Bookings.Include(B => B.CancellationPolicy).FirstOrDefaultAsync(B => B.BookingId == bookingId && B.OwnerId == RequesterId);
-           if (bookingModel == null) {
-            return null;
-           }
-           if (bookingModel.Status != "pending") {
-            throw new Exception("You cannot delete the booking");
-           }
+            var bookingModel = await _dbContext.Bookings.Include(B => B.CancellationPolicy).FirstOrDefaultAsync(B => B.BookingId == bookingId && B.OwnerId == RequesterId);
+            if (bookingModel == null || bookingModel.ServiceId == null)
+            {
+                return null;
+            }
+            if (bookingModel.Status != "pending")
+            {
+                throw new Exception("You cannot delete the booking");
+            }
             _dbContext.CancellationPolicies.Remove(bookingModel.CancellationPolicy);
             _dbContext.Bookings.Remove(bookingModel);
             await _dbContext.SaveChangesAsync();
@@ -51,62 +54,62 @@ namespace api.Repositories
         public async Task<List<Booking>?> GetAllAsync(string role, string RequesterId)
         {
             if (role == "renter")
-            return await _dbContext.Bookings.Include(B => B.Service)
-                                                         .Include(B => B.CancellationPolicy)
-                                                         .Include(B => B.Owner)
-                                                         .Include(B => B.Renter)
-                                                         .Where(B => B.RenterId == RequesterId)
-                                                         .ToListAsync();
-            else if (role == "owner")            
-            return await _dbContext.Bookings.Include(B => B.Service)
-                                                         .Include(B => B.CancellationPolicy)
-                                                         .Include(B => B.Owner)
-                                                         .Include(B => B.Renter)
-                                                         .Where(B => B.OwnerId == RequesterId)
-                                                         .ToListAsync();
-            
+                return await _dbContext.Bookings.Include(B => B.Service)
+                                                             .Include(B => B.CancellationPolicy)
+                                                             .Include(B => B.Owner)
+                                                             .Include(B => B.Renter)
+                                                             .Where(B => B.RenterId == RequesterId && B.ServiceId != null)
+                                                             .ToListAsync();
+            else if (role == "owner")
+                return await _dbContext.Bookings.Include(B => B.Service)
+                                                             .Include(B => B.CancellationPolicy)
+                                                             .Include(B => B.Owner)
+                                                             .Include(B => B.Renter)
+                                                             .Where(B => B.OwnerId == RequesterId && B.ServiceId != null)
+                                                             .ToListAsync();
+
             return null;
-            
+
         }
 
         public async Task<Booking?> GetByIdAsync(string role, int bookingId, string RequesterId)
         {
             if (role.ToLower() == "renter")
-            return await _dbContext.Bookings.Include(B => B.CancellationPolicy)
-                                                        .Include(B => B.Service)
-                                                        .Include(B => B.Renter)
-                                                        .Include(B => B.Owner)
-                                                        .Where(B => B.RenterId == RequesterId)
-                                                        .FirstOrDefaultAsync(B => B.BookingId == bookingId);
+                return await _dbContext.Bookings.Include(B => B.CancellationPolicy)
+                                                            .Include(B => B.Service)
+                                                            .Include(B => B.Renter)
+                                                            .Include(B => B.Owner)
+                                                            .Where(B => B.RenterId == RequesterId)
+                                                            .FirstOrDefaultAsync(B => B.BookingId == bookingId && B.ServiceId != null);
             else if (role.ToLower() == "owner")
-            return await _dbContext.Bookings.Include(B => B.CancellationPolicy)
-                                                        .Include(B => B.Service)
-                                                        .Include(B => B.Renter)
-                                                        .Include(B => B.Owner)
-                                                        .Where(B => B.OwnerId == RequesterId)
-                                                        .FirstOrDefaultAsync(B => B.BookingId == bookingId);
-            
+                return await _dbContext.Bookings.Include(B => B.CancellationPolicy)
+                                                            .Include(B => B.Service)
+                                                            .Include(B => B.Renter)
+                                                            .Include(B => B.Owner)
+                                                            .Where(B => B.OwnerId == RequesterId)
+                                                            .FirstOrDefaultAsync(B => B.BookingId == bookingId && B.ServiceId != null);
+
             return null;
-     
-            
+
+
         }
         public async Task<Booking?> UpdateAsync(string role, int bookingId, Booking updateModel, string requesterId)
         {
             var booking = await _dbContext.Bookings.Include(b => b.CancellationPolicy)
-                                                    .Include(B => B.Service)                                                        .Include(B => B.Service)
+                                                    .Include(B => B.Service)
                                                     .Include(B => B.Renter)
                                                     .Include(B => B.Owner)
-                                                    .Where(b => b.BookingId == bookingId && (b.OwnerId == requesterId || b.RenterId == requesterId))
+                                                    .Where(b => b.BookingId == bookingId && (b.OwnerId == requesterId || b.RenterId == requesterId) && b.ServiceId != null)
                                                     .FirstOrDefaultAsync();
-            if (booking == null) 
+            if (booking == null)
             {
                 throw new Exception("Booking is not found");
-                
+
             }
 
-            if (role.ToLower() == "owner" && booking.OwnerId == requesterId) 
+            if (role.ToLower() == "owner" && booking.OwnerId == requesterId)
             {
-                if (booking.Status.ToLower() == "pending") 
+                if (booking.Status.ToLower() == "pending")
                 {
                     booking.StartDate = updateModel.StartDate;
                     booking.EndDate = updateModel.EndDate;
@@ -118,27 +121,30 @@ namespace api.Repositories
                     await _dbContext.SaveChangesAsync();
                 }
 
-                else if (booking.Status.ToLower() == "booked" && booking.PickUpCode == updateModel.PickUpCode) 
+                else if (booking.Status.ToLower() == "booked" && booking.PickUpCode == updateModel.PickUpCode)
                 {
                     booking.PickUpDate = DateTime.Now;
                     booking.Status = "in-use";
                     booking.ReturnCode = new Random().Next(1000, 9999);
                     await _dbContext.SaveChangesAsync();
                 }
-                else {
+                else
+                {
                     throw new Exception("either the pick up code is not correct or you are not allowed to make the change");
                 }
             }
 
             else if (role.ToLower() == "renter" && booking.RenterId == requesterId)
             {
-                if (booking.Status.ToLower() == "pending" && updateModel.Status.ToLower() == "booked") {
+                if (booking.Status.ToLower() == "pending" && updateModel.Status.ToLower() == "booked")
+                {
                     booking.Status = "booked";
                     booking.CreatedAt = DateTime.Now;
                     booking.PickUpCode = new Random().Next(1000, 9999);
                     await _dbContext.SaveChangesAsync();
                 }
-                else if (booking.Status.ToLower() == "in-use" && booking.ReturnCode == updateModel.ReturnCode) {
+                else if (booking.Status.ToLower() == "in-use" && booking.ReturnCode == updateModel.ReturnCode)
+                {
                     booking.ReturnDate = DateTime.Now;
                     booking.Status = "completed";
                     await _dbContext.SaveChangesAsync();
@@ -156,11 +162,13 @@ namespace api.Repositories
                         throw new Exception("Due to the cancellation policy stated, you cannot cancel");
                     }
                 }
-                else {
+                else
+                {
                     throw new Exception("It is not allowed");
                 }
             }
-            else {
+            else
+            {
                 return null;
             }
